requirements.txt
torch>=2.1
torchvision>=0.16
segmentation-models-pytorch>=0.3.3
albumentations>=1.4
numpy>=1.24
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.65
matplotlib>=3.7
pandas>=2.0
